refactor: remove unusual code
