+++ packages/gravity-vue/src/components/icon/utils.ts
@@ -1,4 +1,5 @@
 import type {
+  SVGDementions,
   SVGIconComponentData,
   SVGIconData,
   SVGIconSpriteData,
@@ -15,7 +16,7 @@ export function isSvgrData(data: SVGIconData): data is SVGIconSvgrData {
 }
 
 export function isComponentSvgData(data: SVGIconData): data is SVGIconComponentData {
-  return (typeof data === 'object' || typeof data === 'function') && 'defaultProps' in data
+  return (typeof data === 'object' || typeof data === 'function')
 }
 
 export function isStringSvgData(data: SVGIconData): data is SVGIconStringData {
@@ -34,4 +35,32 @@ export function getStringViewBox(data: SVGIconStringData) {
   const match = data.match(/viewBox=(["']?)([\d\s,-]+)\1/)
 
   return match ? match[2] : undefined
+}
+
+export function getDementions({ width, height, size, data }: SVGDementions & { data: SVGIconData }) {
+  let w = size || width
+  let h = size || height
+
+  let viewBox: string | undefined
+
+  if (isSpriteData(data)) {
+    viewBox = data.viewBox
+  }
+
+  else if (isComponentSvgData(data)) {
+    w = w || '1em'
+    h = h || '1em'
+  }
+
+  else if (isStringSvgData(data)) {
+    viewBox = getStringViewBox(data)
+  }
+
+  if (viewBox && (!w || !h)) {
+    const values = viewBox.split(/\s+|\s*,\s*/)
+    if (!w) w = values[2]
+    if (!h) h = values[3]
+  }
+
+  return { width: w, height: h }
 }
